fix minus
