+++ paimon-hive/paimon-hive-connector-common/src/main/java/org/apache/paimon/hive/mapred/PaimonInputFormat.java
@@ -19,7 +19,7 @@
 package org.apache.paimon.hive.mapred;
 
 import org.apache.paimon.hive.RowDataContainer;
-import org.apache.paimon.hive.utils.HiveSplitUtils;
+import org.apache.paimon.hive.utils.HiveSplitGenerator;
 import org.apache.paimon.table.FileStoreTable;
 
 import org.apache.hadoop.mapred.InputFormat;
@@ -30,6 +30,7 @@ import org.apache.hadoop.mapred.Reporter;
 
 import java.io.IOException;
 
+import static org.apache.paimon.hive.mapred.PaimonRecordReader.createRecordReader;
 import static org.apache.paimon.hive.utils.HiveUtils.createFileStoreTable;
 
 /**
@@ -41,7 +42,7 @@ public class PaimonInputFormat implements InputFormat<Void, RowDataContainer> {
     @Override
     public InputSplit[] getSplits(JobConf jobConf, int numSplits) {
         FileStoreTable table = createFileStoreTable(jobConf);
-        return HiveSplitUtils.generateSplits(table, jobConf);
+        return HiveSplitGenerator.generateSplits(table, jobConf);
     }
 
     @Override
@@ -49,6 +50,6 @@ public class PaimonInputFormat implements InputFormat<Void, RowDataContainer> {
             InputSplit inputSplit, JobConf jobConf, Reporter reporter) throws IOException {
         FileStoreTable table = createFileStoreTable(jobConf);
         PaimonInputSplit split = (PaimonInputSplit) inputSplit;
-        return HiveSplitUtils.createRecordReader(table, split, jobConf);
+        return createRecordReader(table, split, jobConf);
     }
 }
