+++ paimon-hive/paimon-hive-connector-common/src/main/java/org/apache/paimon/hive/mapred/PaimonRecordReader.java
@@ -25,20 +25,29 @@ import org.apache.paimon.data.InternalRow;
 import org.apache.paimon.data.JoinedRow;
 import org.apache.paimon.hive.RowDataContainer;
 import org.apache.paimon.reader.RecordReaderIterator;
+import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.table.source.ReadBuilder;
 import org.apache.paimon.utils.ProjectedRow;
 
+import org.apache.hadoop.hive.metastore.api.hive_metastoreConstants;
+import org.apache.hadoop.hive.ql.io.IOConstants;
+import org.apache.hadoop.hive.serde2.ColumnProjectionUtils;
+import org.apache.hadoop.hive.serde2.SerDeUtils;
+import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.RecordReader;
 
 import javax.annotation.Nullable;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import static org.apache.paimon.hive.utils.HiveUtils.createPredicate;
 import static org.apache.paimon.hive.utils.HiveUtils.extractTagName;
 
 /**
@@ -162,5 +171,44 @@ public class PaimonRecordReader implements RecordReader<Void, RowDataContainer> 
         // only when the reading finishes will this be set to 1
         // TODO make this more precise
         return progress;
+    }
+
+    public static RecordReader<Void, RowDataContainer> createRecordReader(
+            FileStoreTable table, PaimonInputSplit split, JobConf jobConf) throws IOException {
+        ReadBuilder readBuilder = table.newReadBuilder();
+        createPredicate(table.schema(), jobConf, true).ifPresent(readBuilder::withFilter);
+        List<String> paimonColumns = table.schema().fieldNames();
+        return new PaimonRecordReader(
+                readBuilder,
+                split,
+                paimonColumns,
+                getHiveColumns(jobConf).orElse(paimonColumns),
+                Arrays.asList(getSelectedColumns(jobConf)),
+                table.coreOptions().tagToPartitionField());
+    }
+
+    private static Optional<List<String>> getHiveColumns(JobConf jobConf) {
+        String columns = jobConf.get(IOConstants.SCHEMA_EVOLUTION_COLUMNS);
+        if (columns == null) {
+            columns = jobConf.get(hive_metastoreConstants.META_TABLE_COLUMNS);
+        }
+        String delimiter =
+                jobConf.get(
+                        // serdeConstants.COLUMN_NAME_DELIMITER is not defined in earlier Hive
+                        // versions, so we use a constant string instead
+                        "column.name.delimite", String.valueOf(SerDeUtils.COMMA));
+        if (columns == null || delimiter == null) {
+            return Optional.empty();
+        } else {
+            return Optional.of(Arrays.asList(columns.split(delimiter)));
+        }
+    }
+
+    private static String[] getSelectedColumns(JobConf jobConf) {
+        // when using tez engine or when same table is joined multiple times,
+        // it is possible that some selected columns are duplicated
+        return Arrays.stream(ColumnProjectionUtils.getReadColumnNames(jobConf))
+                .distinct()
+                .toArray(String[]::new);
     }
 }
